+++ start.mjs
@@ -18,13 +18,9 @@
 import path from 'path';
 import { fileURLToPath } from 'url';
 
-import { createRequire } from 'module';
-import { statSync } from 'fs';
-import { execSync } from 'child_process';
 import {
   createRestrictedFs,
   log,
-  runNpm,
   runNpmAndWait,
   runNpx,
   printHeader,
@@ -37,10 +33,8 @@ import {
   prompt,
   killProcessTree,
   sleep,
+  launchDockerContainers,
 } from './scripts/launcher-utils.mjs';
-
-const require = createRequire(import.meta.url);
-const crossSpawn = require('cross-spawn');
 
 const __filename = fileURLToPath(import.meta.url);
 const __dirname = path.dirname(__filename);
@@ -55,89 +49,6 @@ const processes = {
   electron: null,
   docker: null,
 };
-
-/**
- * Sanitize a project name to be safe for use in shell commands and Docker image names.
- * Converts to lowercase and removes any characters that aren't alphanumeric.
- */
-function sanitizeProjectName(name) {
-  return name.toLowerCase().replace(/[^a-z0-9]/g, '');
-}
-
-/**
- * Check if Docker images need to be rebuilt based on Dockerfile or package.json changes
- */
-function shouldRebuildDockerImages() {
-  try {
-    const dockerfilePath = path.join(__dirname, 'Dockerfile');
-    const packageJsonPath = path.join(__dirname, 'package.json');
-
-    // Get modification times of source files
-    const dockerfileMtime = statSync(dockerfilePath).mtimeMs;
-    const packageJsonMtime = statSync(packageJsonPath).mtimeMs;
-    const latestSourceMtime = Math.max(dockerfileMtime, packageJsonMtime);
-
-    // Get project name from docker-compose config, falling back to directory name
-    let projectName;
-    try {
-      const composeConfig = execSync('docker compose config --format json', {
-        encoding: 'utf-8',
-        cwd: __dirname,
-      });
-      const config = JSON.parse(composeConfig);
-      projectName = config.name;
-    } catch (error) {
-      // Fallback handled below
-    }
-
-    // Sanitize project name (whether from config or fallback)
-    // This prevents command injection and ensures valid Docker image names
-    const sanitizedProjectName = sanitizeProjectName(
-      projectName || path.basename(__dirname)
-    );
-    const serverImageName = `${sanitizedProjectName}_server`;
-    const uiImageName = `${sanitizedProjectName}_ui`;
-
-    // Check if images exist and get their creation times
-    let needsRebuild = false;
-
-    try {
-      // Check server image
-      const serverImageInfo = execSync(
-        `docker image inspect ${serverImageName} --format "{{.Created}}" 2>/dev/null || echo ""`,
-        { encoding: 'utf-8', cwd: __dirname }
-      ).trim();
-
-      // Check UI image
-      const uiImageInfo = execSync(
-        `docker image inspect ${uiImageName} --format "{{.Created}}" 2>/dev/null || echo ""`,
-        { encoding: 'utf-8', cwd: __dirname }
-      ).trim();
-
-      // If either image doesn't exist, we need to rebuild
-      if (!serverImageInfo || !uiImageInfo) {
-        return true;
-      }
-
-      // Parse image creation times (ISO 8601 format)
-      const serverCreated = new Date(serverImageInfo).getTime();
-      const uiCreated = new Date(uiImageInfo).getTime();
-      const oldestImageTime = Math.min(serverCreated, uiCreated);
-
-      // If source files are newer than images, rebuild
-      needsRebuild = latestSourceMtime > oldestImageTime;
-    } catch (error) {
-      // If images don't exist or inspect fails, rebuild
-      needsRebuild = true;
-    }
-
-    return needsRebuild;
-  } catch (error) {
-    // If we can't check, err on the side of rebuilding
-    log('Could not check Docker image status, will rebuild to be safe', 'yellow');
-    return true;
-  }
-}
 
 /**
  * Build all production artifacts
@@ -315,47 +226,7 @@ async function main() {
       break;
     } else if (choice === '3') {
       console.log('');
-      log('Launching Docker Container (Isolated Mode)...', 'blue');
-
-      // Check if Dockerfile or package.json changed and rebuild if needed
-      const needsRebuild = shouldRebuildDockerImages();
-      const buildFlag = needsRebuild ? ['--build'] : [];
-
-      if (needsRebuild) {
-        log('Dockerfile or package.json changed - rebuilding images...', 'yellow');
-      } else {
-        log('Starting Docker containers...', 'yellow');
-      }
-      console.log('');
-
-      // Check if ANTHROPIC_API_KEY is set
-      if (!process.env.ANTHROPIC_API_KEY) {
-        log('Warning: ANTHROPIC_API_KEY environment variable is not set.', 'yellow');
-        log('The server will require an API key to function.', 'yellow');
-        log('Set it with: export ANTHROPIC_API_KEY=your-key', 'yellow');
-        console.log('');
-      }
-
-      // Start containers with docker-compose
-      // Will rebuild if Dockerfile or package.json changed
-      processes.docker = crossSpawn('docker', ['compose', 'up', ...buildFlag], {
-        stdio: 'inherit',
-        cwd: __dirname,
-        env: {
-          ...process.env,
-        },
-      });
-
-      log('Docker containers starting...', 'blue');
-      log('UI will be available at: http://localhost:3007', 'green');
-      log('API will be available at: http://localhost:3008', 'green');
-      console.log('');
-      log('Press Ctrl+C to stop the containers.', 'yellow');
-
-      await new Promise((resolve) => {
-        processes.docker.on('close', resolve);
-      });
-
+      await launchDockerContainers({ baseDir: __dirname, processes });
       break;
     } else {
       log('Invalid choice. Please enter 1, 2, or 3.', 'red');
